requirements.txt
# Core PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.4

# NLP and Text Processing
langchain==0.1.0
langchain-community==0.0.10
tiktoken==0.5.2
sentence-transformers==2.2.2
spacy==3.7.2

# OpenAI for Q&A generation
openai==1.6.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# CLI and Configuration
click==8.1.7
python-dotenv==1.0.0
pydantic==2.5.3
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
colorlog==6.8.0
